Fix belongsTo error without inverse relationship
